+++ Runtime/Internal/Gray8Preprocessor.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Buffers;
 
-namespace S46129.QRCode.Internal
+namespace ZXingCpp.QRCode.Internal
 {
     internal static class Gray8Preprocessor
     {